Some changes suggested in rider and rider config files

+++ Infrastructure/deploy.ps1
@@ -25,53 +25,40 @@ Write-Host @"
 az group create --resource-group $ResourceGroupName --location $Location
 
 if ($AppRegistration) {
+    Write-Host "👀 Creating App Registration"
     $App = az ad app create --display-name "OpenVeer-$($Environment)" | Out-string | ConvertFrom-Json
-    $ServicePrinciple = az ad sp create --id $App.appId | Out-string | ConvertFrom-Json
-    $RoleAssignment = az role assignment create --role contributor --subscription $SubscriptionId --assignee-object-id $ServicePrinciple.Id --assignee-principal-type ServicePrincipal --scope "/subscriptions/$($SubscriptionId)/resourceGroups/$($ResourceGroupName)"
-    $Subject = "repo:DevStarOps/OpenVeer:environment:$($Environment):ref:refs/heads/main"
-    if ($Environment -ne "prod") {
-        $Subject = "repo:DevStarOps/OpenVeer:environment:$($Environment)"
+    Write-Host "👀 Creating Service Principal"
+    $SPs = az ad sp list --display-name "OpenVeer-$($Environment)" | Out-string | ConvertFrom-Json
+    if ($SPs.count -eq 0) {
+        $ServicePrinciple = az ad sp create --id $App.appId | Out-string | ConvertFrom-Json
+    } else {
+        $ServicePrinciple = $SPs[0]
     }
-    $Credential = @{
-        "name"="GitHub"
-        "issuer"="https://token.actions.githubusercontent.com"
-        "subject"="$($Subject)"
-        "description"="Deployments for DevStarOps/OpenVeer"
-        "audiences"=@("api://AzureADTokenExchange")
+    Write-Host "👀 Creating Role Assignment"
+    $RoleAssignments = az role assignment list --scope "/subscriptions/$($SubscriptionId)/resourceGroups/$($ResourceGroupName)" --assignee $App.appId
+    if ($RoleAssignments.count -eq 0) {
+        $RoleAssignment = az role assignment create --role contributor --subscription $SubscriptionId --assignee-object-id $ServicePrinciple.Id --assignee-principal-type ServicePrincipal --scope "/subscriptions/$($SubscriptionId)/resourceGroups/$($ResourceGroupName)"
     }
-    $TempBodyFile = "$([System.IO.Path]::GetTempFileName()).json"
-    Set-Content -Path $TempBodyFile -Value ($Credential | ConvertTo-Json) -Encoding utf8
-    $federatedCredential = az ad app federated-credential create --id $App.Id --parameters $TempBodyFile
-    Remove-Item -LiteralPath $TempBodyFile
+    Write-Host "👀 Creating Federated Credential"
+    $FederatedCredentials = az ad app federated-credential list --id $App.Id
+    if ($FederatedCredentials.count -eq 0) {
+        $Subject = "repo:DevStarOps/OpenVeer:environment:$($Environment):ref:refs/heads/main"
+        if ($Environment -ne "prod") {
+            $Subject = "repo:DevStarOps/OpenVeer:environment:$($Environment)"
+        }
+        $Credential = @{
+            "name"="GitHub"
+            "issuer"="https://token.actions.githubusercontent.com"
+            "subject"="$($Subject)"
+            "description"="Deployments for DevStarOps/OpenVeer"
+            "audiences"=@("api://AzureADTokenExchange")
+        }
+        $TempBodyFile = "$([System.IO.Path]::GetTempFileName()).json"
+        Set-Content -Path $TempBodyFile -Value ($Credential | ConvertTo-Json) -Encoding utf8
+        $federatedCredential = az ad app federated-credential create --id $App.Id --parameters $TempBodyFile
+        Remove-Item -LiteralPath $TempBodyFile
+    }
 }
 
-Write-Host @"
-
-👀 Deploy Core App
-
-"@
-
+Write-Host "👀 Deploy Core App"
 az deployment group create --resource-group $ResourceGroupName --template-file "core-app.bicep" --parameters "environments/core-app-$($Environment).json" sqlAdministratorLoginPassword=$SqlAdminPassword
-
-Write-Host @"
-
-👀 Remove extra resources
-
-"@
-
-az deployment group create --force "core-app.bicep" --resource-group $ResourceGroupName --mode Complete
-
-
-
-$MyJsonHashTable = @{
-    'MyList' = @{
-      'Item1' = @{
-        'Name' = 'AMD Ryzen 5 3600x'
-        'Type' = 'CPU'
-        'Price' = '$69.99'
-        'Where' = 'Amazon.com'
-      }
-    }
-  }
-  
-  $MyJsonVariable = $MyJsonHashTable | ConvertTo-Json
